Move bet types
